Add project files.

+++ Functions/main.cpp
@@ -305,9 +305,9 @@ int Mul(int num_1, int num_2)
     return num_1 * num_2;
 }
 
-int Div(int num_1, int num_2) 
+double Div(int num_1, int num_2) 
 {
-    return num_1 / num_2;
+    return (double)num_1 / num_2;
 }
 
 void main()
